refactoring images into two folders
