+++ studyhelper/src/Components/Body/Body.css
@@ -40,7 +40,7 @@
   padding-top: 3em;
   padding-bottom: 3em;
   justify-content: space-between;
-  border-bottom: 1px solid #9199c6;
+  border-bottom: 1px solid #9da2c0;
 }
 
 .content-item div img {
@@ -77,7 +77,7 @@ footer {
 }
 
 footer ul {
-  padding: 1em 0 1em 1em;
+  padding: 0.5em 0 0.5em 1em;
   margin-bottom: 0em;
 }
