+++ studyhelper/src/Components/NavBar/NavBar.jsx
@@ -5,9 +5,15 @@ const NavBar = () => {
   return (
     <>
       <ul>
-        <li>Home</li>
-        <li>Contact Us</li>
-        <li>Login</li>
+        <li>
+          <a href='https://www.google.com/'>Home</a>
+        </li>
+        <li>
+          <a href='https://www.google.com/'>Contact Us</a>
+        </li>
+        <li>
+          <a href='https://www.google.com/'>Login</a>
+        </li>
       </ul>
     </>
   );
