+++ studyhelper/src/Pages/Homepage/Homepage.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../Homepage/Homepage.css';
-import { Cover } from '../../Components/Cover/Cover';
-import { Body } from '../../Components/Body/Body';
+import Cover from '../../Components/Cover/Cover';
+import Body from '../../Components/Body/Body';
 
 const Homepage = () => {
   return (
