+++ studyhelper/src/index.js
@@ -1,7 +1,6 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import './index.css';
-import 'bootstrap/dist/css/bootstrap.min.css';
 import Router from './Routes'
 import NavBar from './Pages/components/NavBar/NavBar';
